* info/nodes.c (info_create_tag), info/nodes.h (TAG): set TAG nodestart to be size_t and initialize to 0.  Indeed, right after info_create_tag is called, the tag nodestart is set to a non-negative value.  It should not be possible to have 0 for nodestart from build_tag_table as there is at least the node separator.  A bogus entry in tag table or a 0 in tag table is possible and although it is unlikely to be correct, nodestart can be adjusted.
* info/nodes.c (adjust_nodestart, info_node_of_tag_ext): remove
conditions on tag->nodestart being >= 0, as it is always true.

+++ info/nodes.h
@@ -102,7 +102,7 @@ typedef struct {
 typedef struct {
   char *filename;               /* The file where this node can be found. */
   char *nodename;               /* The node pointed to by this tag. */
-  long nodestart;               /* The value read from the tag table. */
+  size_t nodestart;             /* The value read from the tag table. */
   long nodestart_adjusted;      /* Where the node or anchor actually is. */
   int flags;                    /* Same as NODE.flags. */
   NODE cache;                   /* Saved information about pointed-to node. */
